#326: move ScopedJob to util, trim KDoc, make stop-cancels test non-vacuous
