+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/discovery/DiscoveredDevicesStore.kt
@@ -1,6 +1,7 @@
 package com.tubetoast.tether.discovery
 
 import com.tubetoast.tether.protocol.Device
+import com.tubetoast.tether.util.ScopedJob
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.ExperimentalForInheritanceCoroutinesApi
 import kotlinx.coroutines.delay
