+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/discovery/RendezvousAnnouncer.kt
@@ -2,6 +2,7 @@ package com.tubetoast.tether.discovery
 
 import com.tubetoast.tether.network.FileClient
 import com.tubetoast.tether.protocol.Device
+import com.tubetoast.tether.util.ScopedJob
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.update
